+++ src/engine/client/client.h
@@ -21,6 +21,7 @@ PL_EXTERN_C
 
 void Client_Initialize(void);
 void Client_Simulate(void);
+void Client_Render(double delta);
 void Client_Shutdown(void);
 
 PL_EXTERN_C_END
